代码优化

+++ dlz.framework/src/main/java/com/dlz/framework/redis/excutor/IJedisStringExecutor.java
@@ -127,7 +127,7 @@ public interface IJedisStringExecutor extends IJedisExecutor {
             byte[] key1 = SafeEncoder.encode(JedisKeyUtils.getRedisKey(key));
             j.set(key1, SerializeUtil.serialize(value));
             if (seconds > 0) {
-                j.expire(key1, seconds);
+                j.expire(key1, (long)seconds);
             }
             return true;
         });
